Rename remote_data module to sync
